Chore: moved test location
