+++ jest.config.js
@@ -1,10 +1,10 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
   testTimeout: 10000,
-  preset: "ts-jest",
-  moduleFileExtensions: ["js", "json", "ts"],
-  testEnvironment: "node",
+  preset: 'ts-jest',
+  moduleFileExtensions: ['js', 'json', 'ts'],
+  testEnvironment: 'node',
   transform: {
-    "^.+\\.(t|j)s$": "ts-jest",
+    '^.+\\.test\\.(t|j)s$': 'ts-jest',
   },
 };
