+++ tsconfig.json
@@ -6,6 +6,7 @@
     "noUnusedLocals": false,
     "removeComments": false,
     "noLib": false,
+    "skipLibCheck": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "target": "es6",
@@ -13,5 +14,5 @@
     "outDir": "./dist"
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist/**/*"]
 }
